Update queue test class
